showing the new password
