+++ public/assets/index.html
@@ -12,7 +12,7 @@
     <div class="main__container">
         <h1>Password Generator</h1>
         <div class="password-generator-form">
-            <h2>Select the options</h2>
+            <h2 class="headling2">Select the options</h2>
             <div>
                 <label for="">Number of Characters</label>
                 <input type="number" name="number-of-characters" id="number-of-characters" max="16">
