+++ src/assets/css/styles.css
@@ -41,7 +41,7 @@ body {
     }
 }
 
-.password-generator-form h2 {
+.headling2 {
     color: rgb(112, 45, 175);
     font-size: 45px;
     margin-bottom: 30px;
